Rethink everything again
After reading over the messaging example in goka, it's clear that
my mental picture of what APIs are available in it and Kafka
Streams is still very wrong.  Previously I thought that the
'sstable' indexes that are used for 'local state' are themselves
stored in kafka topics; however that doesn't seem to be the way
it works.  It looks like you can do 1:1 key maps and joins pretty
easily, and all leveldb stuff in Goka is local disk storage.  So,
adjust the design to deal with this.  A sorted string table on a
topic with a low recovery time would be cool, but it looks like
it is not to be.

There are also tangible design changes, with a clear distinction
being drawn between components which are doing live streaming
(which, of course, need to all emit single match rows, not groups
of matches for low latency), semi-live streaming (the windowed
join in grep-cursor) and deep search (grep-logs).

Also throw in some thoughts about how to implement progress
updates and a 'count' function (or timeseries statistics).